requirements.txt
matplotlib
numpy
pandas
Pillow
tqdm
